+++ apps/pushnotif-func/src/functions/HandleNHDeleteInstallationCallActivity/handler.ts
@@ -31,24 +31,15 @@ export { ActivityResultSuccess } from "../../utils/durable/activities";
 export const getActivityBody =
   (
     nhPartitionFactory: NotificationHubPartitionFactory,
-    nhNewPartitionFactory: NotificationHubPartitionFactory,
     telemetryClient: TelemetryClient,
     // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
   ): ActivityBody<ActivityInput, ActivityResultSuccess> =>
   ({ input, logger }) => {
     logger.info(`INSTALLATION_ID=${input.installationId}`);
     const nhClient = nhPartitionFactory.getPartition(input.installationId);
-    const nhLegacyClient = nhNewPartitionFactory.getPartition(
-      input.installationId,
-    );
 
     return pipe(
-      deleteInstallation(
-        nhClient,
-        nhLegacyClient,
-        input.installationId,
-        telemetryClient,
-      ),
+      deleteInstallation(nhClient, input.installationId),
       TE.bimap(
         (e) => {
           telemetryClient.trackEvent({