+++ apps/pushnotif-func/src/functions/HandleNHDeleteInstallationCallActivity/index.ts
@@ -22,10 +22,6 @@ const nhPatitionFactory = new NotificationHubPartitionFactory(
   config.AZURE_NOTIFICATION_HUB_PARTITIONS,
 );
 
-const nhNewPartitionFactory = new NotificationHubPartitionFactory(
-  config.AZURE_NEW_NOTIFICATION_HUB_PARTITIONS,
-);
-
 /**
  * Build a `HandleNHDeleteInstallationCallActivity` to be called by an Orchestrator
  *
@@ -45,7 +41,7 @@ const activityFunctionHandler = createActivity<ActivityInput>(
   activityName,
   ActivityInput,
   ActivityResultSuccess,
-  getActivityBody(nhPatitionFactory, nhNewPartitionFactory, telemetryClient),
+  getActivityBody(nhPatitionFactory, telemetryClient),
 );
 
 export default activityFunctionHandler;