+++ apps/send-func/src/adapters/functions/aar-notifications.ts
@@ -1,19 +1,21 @@
 import { iunSchema, mandateIdSchema } from "@/domain/notification.js";
+import { TelemetryEventName, TelemetryService } from "@/domain/telemetry.js";
 import { GetNotificationUseCase } from "@/domain/use-cases/get-notification.js";
 import { HttpRequest, InvocationContext } from "@azure/functions";
 import { LollipopHeaders } from "io-messages-common/adapters/lollipop/definitions/lollipop-headers";
 import { ExtentedHttpHandler } from "io-messages-common/adapters/middleware";
 
-import {
-  AarGetNotificationResponse,
-  aarProblemJsonSchema,
-} from "../send/definitions.js";
+import { AarGetNotificationResponse } from "../send/definitions.js";
 import { NotificationClientError } from "../send/notification.js";
-import { malformedBodyResponse } from "./commons/response.js";
+import {
+  malformedBodyResponse,
+  sendProblemToAARProblemJson,
+} from "./commons/response.js";
 
 export const getNotification =
   (
     getNotificationUseCase: GetNotificationUseCase,
+    telemetryService: TelemetryService,
   ): ExtentedHttpHandler<LollipopHeaders> =>
   async (
     request: HttpRequest,
@@ -58,8 +60,15 @@ export const getNotification =
       if (err instanceof NotificationClientError) {
         context.error("Notification client error:", err.message);
 
+        telemetryService.trackEvent(
+          TelemetryEventName.SEND_AAR_NOTIFICATION_SERVER_ERROR,
+          {
+            status: err.status,
+          },
+        );
+
         return {
-          jsonBody: aarProblemJsonSchema.parse(err.body),
+          jsonBody: sendProblemToAARProblemJson(err.body),
           status: 500,
         };
       }
@@ -67,6 +76,7 @@ export const getNotification =
       const errorMessage =
         err instanceof Error ? err.message : JSON.stringify(err);
       context.error(err);
+
       return {
         jsonBody: {
           detail: errorMessage,