+++ apps/send-func/src/adapters/functions/aar-qrcode-check.ts
@@ -1,3 +1,4 @@
+import { TelemetryEventName, TelemetryService } from "@/domain/telemetry.js";
 import { QrCodeCheckUseCase } from "@/domain/use-cases/qr-code-check.js";
 import { HttpRequest, InvocationContext } from "@azure/functions";
 import { LollipopHeaders } from "io-messages-common/adapters/lollipop/definitions/lollipop-headers";
@@ -5,18 +6,21 @@ import { ExtentedHttpHandler } from "io-messages-common/adapters/middleware";
 
 import {
   AarQRCodeCheckResponse,
-  aarProblemJsonSchema,
   checkQrMandateRequestSchema,
 } from "../send/definitions.js";
 import {
   NotRecipientClientError,
   NotificationClientError,
 } from "../send/notification.js";
-import { malformedBodyResponse } from "./commons/response.js";
+import {
+  malformedBodyResponse,
+  sendProblemToAARProblemJson,
+} from "./commons/response.js";
 
 export const aarQRCodeCheck =
   (
     qrCodeCheckUseCase: QrCodeCheckUseCase,
+    telemetryService: TelemetryService,
   ): ExtentedHttpHandler<LollipopHeaders> =>
   async (
     request: HttpRequest,
@@ -54,7 +58,7 @@ export const aarQRCodeCheck =
       return { jsonBody: response, status: 200 };
     } catch (err) {
       if (err instanceof NotRecipientClientError) {
-        context.error("NotRecipient client error:", err.message);
+        //SEND returns 403. It will start the user delegation mobile flow
         return {
           jsonBody: err.body,
           status: 403,
@@ -64,8 +68,29 @@ export const aarQRCodeCheck =
       if (err instanceof NotificationClientError) {
         context.error("Notification client error:", err.message);
 
+        switch (err.status) {
+          case 403:
+            telemetryService.trackEvent(
+              TelemetryEventName.SEND_AAR_QRCODE_CHECK_MALFORMED_403,
+            );
+            break;
+          case 404:
+            telemetryService.trackEvent(
+              TelemetryEventName.SEND_AAR_QRCODE_CHECK_DATA_NOT_FOUND,
+            );
+            break;
+          default:
+            telemetryService.trackEvent(
+              TelemetryEventName.SEND_AAR_QRCODE_CHECK_SERVER_ERROR,
+              {
+                status: err.status,
+              },
+            );
+            break;
+        }
+
         return {
-          jsonBody: aarProblemJsonSchema.parse(err.body),
+          jsonBody: sendProblemToAARProblemJson(err.body),
           status: 500,
         };
       }
@@ -73,6 +98,7 @@ export const aarQRCodeCheck =
       const errorMessage =
         err instanceof Error ? err.message : JSON.stringify(err);
       context.error(err);
+
       return {
         jsonBody: {
           detail: errorMessage,